[skip changelog] Document sketch filename requirements (#1287)
Previously, the Arduino Sketch Specification only documented the requirements for sketch folder names. Since the primary
sketch filename must match the folder name, this appeared at a glance to be self-evident. However, sketches may consist
of multiple files, and so the specification was ambiguous regarding the naming requirements for additional files.