[skip changelog] Lib install dir path is now canonical
This stems mainly from the fact the CI uses 8.3 naming convention for
paths on Windows, that causes several issues when trying to verify that
certain paths are printed in the output of certain commands.

Making sure that canonical paths are set as installation directory for
all libraries makes it easier to test.

For more info on 8.3 paths see: https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file#short-vs-long-names