[skip changelog] Document "Library Name Priority" (#1301)
When multiple libraries contain files matching an `#include` directive, Arduino CLI must pick one. Multiple factors are
used in order to make an intelligent determination of which library is best. In order to enhance this determination, the
library.properties `name` value is now one of those factors. The other factors are already documented in the "Sketch
Build Process" documentation, but this newly added feature was previously undocumented.