[skip changelog] Remove redundant artifact upload from test workflow
Previously, there were two steps in the test workflow for the Linux ARM64 artifact upload. Although it did no real harm
since it simply replaced the artifact from the first upload with a duplicate, it does slow down the workflow and adds a
little more of a maintenance burden, so better to remove it.