[skip changelog] Update JSON output key name in integration test helper
The key names used in Arduino CLI's JSON output were made consistent and standardized. One of the helper functions for
the integration tests was not updated accordingly. This was missed because it is only used when running the tests locally
due to requiring a board to be connected to the computer.