[skip changelog] Remove obsolete go generate directive (#1341)
In the early days of Arduino CLI's gRPC interface, the protocol buffer compilation was done via `go generate`, configured
by a directive comment. The `protoc:compile` task is now used instead, but the old directive was left in place, where it
became outdated through changes to the repository structure.

This directive no longer functions or serves a purpose and might cause confusion to contributors. So it should be removed.