[skip changelog] Remove obsolete "Dependabot Preview" configuration file (#1372)
Dependabot Preview shut down 2021-08-03. It is replaced by GitHub-native Dependabot, which uses the
`.github/dependabot.yml` configuration file, with a different data format.

From an investigation into how the obsolete configuration could be migrated to the new format, I determined that the
specific "security updates only" configuration used by this repository is not provided by the new configuration file,
which is for configuring updates of any outdated dependencies. The "security updates" feature is configured via the
repository settings (Settings > Security & analysis).

Since the old file is no longer used, and might cause confusion, it is best to remove it from the repository.