[skip changelog] Sync general formatting check workflow with template
We have assembled a collection of reusable GitHub Actions workflows:
https://github.com/arduino/tooling-project-assets
These workflows will be used in the repositories of all Arduino tooling projects.

The Prettier formatting tool is used in the tasks for formatting and checking the formatting of the repository's
Markdown and YAML files. Prettier supports a long list of other languages. There is no reason not to take advantage of
those capabilities as well, so the new workflow applies it to all supported files.

Several of the tasks have been deprecated in favor of the new `general:format-prettier` task:

- `docs:check`
- `docs:format`
- `config:check`
- `config:format`

These still work as before, but they now display a deprecation warning.