Remove superfluous formatting tasks
These are replaced by the `general:format-prettier` task. Even though that doesn't provide a formatting compliance check,
the developer doesn't have any need for such a thing, since they will be better off to simply format, and the CI now
takes the approach of formatting and then failing if it produced a diff.