[skip changelog] Add skip changelog prefix to Dependabot workflow update commits (#1381)
Dependabot is configured to submit pull requests for updates to the GitHub Actions actions used in the repository's
workflows whenever a new version is available. These commits will not result in any user facing changes and so should not
be mentioned in the release changelog.

The changelog generation system automatically excludes any commit that has the `[skip changelog]` prefix on its message
title. Adding this changelog to the commits generated by Dependabot will save the release manager time when cleaning up
the raw changelog.