[skip changelog] Sync general formatting check workflow with template (#1380)
* [skip changelog] Sync general formatting check workflow with template

We have assembled a collection of reusable GitHub Actions workflows:
https://github.com/arduino/tooling-project-assets
These workflows will be used in the repositories of all Arduino tooling projects.

The Prettier formatting tool is used in the tasks for formatting and checking the formatting of the repository's
Markdown and YAML files. Prettier supports a long list of other languages. There is no reason not to take advantage of
those capabilities as well, so the new workflow applies it to all supported files.

Several of the tasks have been deprecated in favor of the new `general:format-prettier` task:

- `docs:check`
- `docs:format`
- `config:check`
- `config:format`

These still work as before, but they now display a deprecation warning.

* Remove superfluous formatting tasks

These are replaced by the `general:format-prettier` task. Even though that doesn't provide a formatting compliance check,
the developer doesn't have any need for such a thing, since they will be better off to simply format, and the CI now
takes the approach of formatting and then failing if it produced a diff.

* [skip changelog] Bring all files into Prettier formatting style compliance

Previously, Prettier was only run on the Markdown and YAML files. With the new expanded scope of Prettier coverage, the
JSON files are now required to be compliant.