[skip changelog] Sync documentation assets with templates (#1383)
* [skip changelog] Sync documentation assets with templates

We have assembled a collection of reusable project assets:
https://github.com/arduino/tooling-project-assets
These assets will be used in the repositories of all Arduino tooling projects.

Some significant improvements and standardizations have been made in the upstream "template" assets, and those are
introduced to this repository here.

Notable:

- Use markdownlint to check for common problems and formatting.
- Make link check task Windows compatible
- Streamline website versioning system
- Use modern MkDocs versioning features
- Make command reference generation task Windows compatible
- Consolidate all Python dependency management to Poetry
- Standardize documentation-related task names and deprecate old task names

* Remove obsolete documentation tasks

The sync with the template documentation assets has resulted in new standardized names for some of the documentation
management tasks:

- `docs:gen:commands` -> `go:cli-docs`
- `docs:gen` -> `docs:generate`
- `docs:build` -> `docs:check`
- `docs:serve` -> `website:serve`
- `docs:check-links` -> `markdown:check-links`

* [skip changelog] Use standard list prefix in Markdown

This is the standardized ordered list prefix for use in all Arduino Tooling Markdown files.

The Markdown renderer will automatically handle numbering. Manual numbering means that either all following list items
need to be renumbered every time a list item is added or removed, resulting in more work for contributors and reviewers.

* [skip changelog] Use standardized Markdown code block style

"Code fencing" is superior to the indentation code block syntax of Markdown because it permits syntax highlighting. Even
in cases where syntax highlighting is not applicable, a consistent style should be used throughout all the Markdown files
of Arduino Tooling projects.