[skip changelog] Move website versioning information out of CONTRIBUTING.md (#1386)
* [skip changelog] Fix broken workflow link

The GitHub Actions workflow used to deploy the documentation website was renamed to a filename more suitable for
existence in a collection of reusable workflows that have multiple variations according to project requirements.

The link to the file was not updated accordingly, which resulted it breaking.

* [skip changelog] Move website versioning information out of CONTRIBUTING.md

This content is for maintainers, and is not of interest to the average contributor. It makes an already overly long
contributor guide even more overwhelming, and thus likely to be skipped.

The documentation versioning and publishing system is now maintained in a centralized repository hosting a collection of
reusable project assets. The documentation content previously located in CONTRIBUTING.md has all been migrated there and
expanded.

Since this is not high level contributor guidelines, it will be better to maintain it in a single place, linking to that
documentation as needed.