Remove duplicate task for protobuf docs generation (#1387)
The "gRPC reference" section of the documentation website is automatically generated from the repository's protocol
buffer files. In order to make this generation process easy for developers to run, the necessary commands are defined in
a task.

For some reason, there were two tasks for this purpose, with different names and descriptions, but identical commands.
This makes the taskfile more difficult to understand and more difficult to maintain.

One of the tasks, `protoc:docs` is called by the `docs:generate` task. The other, `docs:gen:protobuf`, is not used or
referenced anywhere in the repository. So I have removed `docs:gen:protobuf` from the taskfile.