[skip changelog] Revert "Add skip changelog prefix to Dependabot workflow update commits (#1381)"
This reverts commit 4fbf7b7b047bbd63c00cd2dcc7f3b0e0551ea7ec.

It turns out there is an undocumented maximum length of 15 characters for this prefix:

https://github.com/arduino/arduino-cli/network/updates

```
Dependabot encountered the following error when parsing your .github/dependabot.yml:

The property '#/updates/0/commit-message/prefix' was not of a maximum string length of 15
Please update the config file to conform with Dependabot's specification.
```
Unfortunately, even though it is possible to reduce the prefix length by 1 with the removal of the trailing space, we
still end up with 16 characters.