[skip changelog] Run formatting and linting CI checks on all Go modules
The repository contains multiple Go modules. Previously, the CI workflow that lints and checks the formatting of the
repository's Go code was only checking the primary module in the root of the repo.

This workflow is now expanded to cover all Go code through the use of a dedicated matrix job for each of the modules.

Arbitrary Go module paths can be specified to the tasks by defining the `GO_MODULE_PATH` environment variable. If this
variable is not defined, the default root module path is used as default, preserving the previous task behavior.