[skip changelog] Use major version refs of action in "Test Go" workflow
Use of the major version ref will cause the workflow to benefit from ongoing development to the actions up until such
time as a new major release of an action is made, at which time we would need to evaluate whether any changes to the
workflow are required by the breaking change that triggered the major release before updating the major ref
(e.g., `uses: codecov/codecov-action@v3`).

The previous pin to the patch version required an update to the workflow on every action release in order to keep it
updated.