Draft: Exclude compile_commands.json from sketch loading
This allows users to create a symlink to the autogenerated file in the
sketch directory, without that file being handled or causing failure
when the target file does not exist yet (or anymore).

Draft: This abuses FilterOutSuffix, it should match the filename, but
go-paths-helper does not support this now.

Draft: Is this really the right place/approach to exclude this file?