[skip changelog] Temporarily full pin Python in integration test CI workflow
A bug introduced in the 3.9.7 release of Python causes a spurious failure of the
`test\test_lib.py::test_install_git_invalid_library` integration test:
https://bugs.python.org/issue45121

As a workaround, the Python version used by the "Test Integration" CI workflow that runs the integration tests is pinned
to the last working version: 3.9.6. Since it is convenient to get automatic updates for Python patch releases, this full
pin should be reverted back to the "3.9" minor version pin once a new version of Python is released with the bug fixed
and added to versions available for installation via the `actions/setup-python` GitHub Actions action.