[skip changelog] Fully document library layouts in specification
Arduino library authors may chose from multiple structures, each of which have different implications for the compilation
of the library source files. The specification explains this in a somewhat simplistic manner as "1.5 format" and
"1.0 format". That is sufficient for most, but does not provide a complete picture. Feedback indicated that more advanced
users were left wanting complete documentation, which is provided here.

The subject is not well suited for description via prose, so a chart and table are employed instead.

The chart is defined in the DOT graph description language.