[skip changelog] Document `packages[*].platforms[*].deprecated` package index field
In the event an Arduino boards platform is deprecated by the maintainer, it will be useful to clearly communicate this
fact to the user and ensure they will easily find a preferred alternative platform. In order to facilitate this, a
`packages[*].platforms[*].deprecated`field was added to the Arduino package index format and this information is used to
enhance the user interfaces of the official Arduino development software.

This useful feature is now documented in the Arduino package index specification.