[skip changelog] Sync install docs from "template" (#1544)
* Add an introduction to install docs

This document provides a list of installation options. The user may pick any one option that is most convenient for them.
However, a reader might get the impression that it is a list of instructions that must be followed from top to bottom.

An introduction makes the situation immediately clear to the reader.

* Clarify that sh is not required in template install docs

A significant portion of Arduino users use Windows. The first option listed in the installation instructions is for an
installation shell script, which can be used on Windows, but might not be convenient for some Windows users.

This could be discouraging or frustrating to potential new users, but not necessarily so because the script is an
optional convenience and directly downloading and installing the tool is also an option, as documented in the
instructions for the second listed option.

A little refinement of the wording of the script's `sh` requirement makes the situation clear to those users.

* Add a reference link re: PATH to install docs

A reference link provides easy access to the relevant information for readers who are not familiar with this concept
while not harming readability for everyone else by increasing verbosity.

* Align heading levels in install docs

This document is intended to provide a list of installation options:

- brew
- script
- download
- build

but the heading levels were misaligned, which resulted in this structure:

-brew
- script
  - download
  - build