[skip changelog] Make use of 'avr' as example platform more consistent (#1680)
* Make use of 'avr' as example platform more consistent

It's fairly easy to miss the one line early in the doc that says this text uses 'avr' as an example, leading a reader to possibly believe the documentation hasn't been updated since 'avr' was the only platform choice. Within the doc, the text was a little inconsistent as to how it referred to the avr platform and platform specific tools. This change tries to standardize how we refer to the example platform.

* Fix a couple lint errors.

* Fix prettier lint errors.

* Fix an additional prettier lint error

* Update docs/sketch-build-process.md

Co-authored-by: per1234 <accounts@perglass.com>

Co-authored-by: per1234 <accounts@perglass.com>