[skip changelog] Remove broken download stats workflow (#1701)
The "arduino-stats" GitHub Actions workflow was designed to periodically gather download statistics from Arduino CDN and
push results to Datadog.

The recorded stats showed a periodic decrease in total download count. Since this is patently impossible, it is clear
that something was wrong with the system and that the recorded data was not usable. An investigation into the problem
was never done.

On 2022-03-14, the runs of the "arduino-stats" GitHub Actions workflow began to fail. Because there had not been any
relevant change in the repository between the last successful run and the first failing run, it seems that some external
change caused the breakage. Since the workflow was not ever working successfully and the lack of an investigation about
that indicates that the stats are not of immediate importance, the best course of action is to simply remove the broken
infrastructure from the repository rather than investing time into fixing something that isn't being used anyway.