[skip changelog] Remove unused GitHub release download stats workflow (#1702)
The "github-stats" GitHub Actions workflow periodically gathers GitHub release asset download statistics for Arduino CLI
and pushes the results to Datadog.

There are no known problems with this workflow. However, the companion "arduino-stats" workflow that did the same for
the downloads of Arduino CLI from downloads.arduino.cc was broken and thus removed from the repository.

The GitHub stats are not very valuable on their own as they only provide an unknown fraction of the total downloads of
Arduino CLI. They have also not ended up being used. Since it doesn't provide any value and represents a maintenance
burden, it is hereby removed from the repository.