Fix installation script bugs affecting Apple M1 users
The script which provides convenient installation of Arduino CLI is hosted and maintained in a separate repository which
contains the collection of reusable tooling project assets. Some defects were present in the script's handling of lack
of a build for a specific host architecture, which manifested when users with Apple M1 systems attempted to use it.

These defects were fixed and some general improvements made to the related code.

- Remove broken frivolous function from install script
  https://github.com/arduino/tooling-project-assets/commit/127116b30163e5c63384d6a46c57f9f5d093d61a
- Remove unnecessary use of eval from install script
  https://github.com/arduino/tooling-project-assets/commit/e2b5740daf890aa20d27d901ee8fc11743b38036
- Fallback to x86-64 release when macOS ARM 64-bit build not available
  https://github.com/arduino/tooling-project-assets/commit/d8c59e3caac6182077e9bbefe3a5209c3ec96782

Those are pulled into the Arduino CLI project here.