[skip changelog] Sync "Check Go Dependencies" workflow with upstream
A "Check Go Dependencies" GitHub Actions workflow is used to check whether all Go package dependencies of the modules in
this repository use compatible licenses.

This workflow is hosted and maintained in a repository dedicated to a collection of such reusable Arduino tooling project assets.

Several enhancements have been made to the upstream file, which are pulled into the Arduino CLI repository here:

- Add schedule event trigger to catch breakage caused by external changes
  https://github.com/arduino/tooling-project-assets/commit/e71e4702279bd165769efd830f276b6b492de92c
- Run workflow on release branch creation
  https://github.com/arduino/tooling-project-assets/commit/22504dc85c1bea37613506016e677cd545a9af1b