Use real names for libs in upgrade commands
The real name for a lib (the name stored in the library.properties)
is required to find a library in the index. The installed name, which
matches the diretory name, is not valid for looking up libraries in
the index.

Libraries that do not have a library.properties file, or where the
name value in that file does not match the value in library_index.json,
cannot be upgrade by the CLI.

Also Fix tests for outdated libs

The tests for update and outdated changed the library.properties
file for an installed lib, but also wiped out the name field.
Since we rely on that field to determine the library name when
using the RealName property, the test's change was causing a failure
outside the scope of the test itself.